LPS-180469 Rename files and replace usages
