+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/components/criteria_builder/DropZone.js
