+++ modules/apps/segments/segments-web/src/main/resources/META-INF/resources/js/components/segment_edit/SegmentEdit.js
@@ -37,7 +37,7 @@ import {
 import {initialContributorShape} from '../../utils/types.es';
 import {sub} from '../../utils/utils';
 import ContributorInputs from '../criteria_builder/ContributorInputs.es';
-import ContributorsBuilder from '../criteria_builder/ContributorsBuilder.es';
+import ContributorsBuilder from '../criteria_builder/ContributorsBuilder';
 import LocalizedInput from '../title_editor/LocalizedInput';
 
 function SegmentEdit({
