+++ modules/apps/segments/segments-web/test/js/components/criteria_builder/CriteriaGroup.es.js
@@ -16,7 +16,7 @@ import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 import {wrapInTestContext} from 'react-dnd-test-utils';
 
-import {default as Component} from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaGroup.es';
+import {default as Component} from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaGroup';
 
 const connectDnd = jest.fn((element) => element);
