+++ modules/apps/segments/segments-web/test/js/components/criteria_builder/CriteriaRow.es.js
@@ -15,7 +15,7 @@
 import {render, waitFor} from '@testing-library/react';
 import React from 'react';
 
-import CriteriaRow from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaRow.es';
+import CriteriaRow from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaRow';
 import {PROPERTY_TYPES} from '../../../../src/main/resources/META-INF/resources/js/utils/constants';
 
 import '@testing-library/jest-dom/extend-expect';
