+++ modules/apps/segments/segments-web/test/js/components/criteria_builder/CriteriaRowReadable.es.js
@@ -15,7 +15,7 @@
 import {render} from '@testing-library/react';
 import React from 'react';
 
-import CriteriaRowReadable from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaRowReadable.es';
+import CriteriaRowReadable from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/CriteriaRowReadable';
 import {dateToInternationalHuman} from '../../../../src/main/resources/META-INF/resources/js/utils/utils';
 
 import '@testing-library/jest-dom/extend-expect';
