+++ modules/apps/segments/segments-web/test/js/components/criteria_builder/DropZone.es.js
@@ -15,7 +15,7 @@
 import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 
-import DropZone from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/DropZone.es';
+import DropZone from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/DropZone';
 
 const connectDnd = jest.fn((element) => element);
