+++ modules/apps/segments/segments-web/test/js/components/criteria_builder/EmptyDropZone.es.js
@@ -15,7 +15,7 @@
 import {cleanup, render} from '@testing-library/react';
 import React from 'react';
 
-import EmptyDropZone from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/EmptyDropZone.es';
+import EmptyDropZone from '../../../../src/main/resources/META-INF/resources/js/components/criteria_builder/EmptyDropZone';
 
 const connectDnd = jest.fn((element) => element);
